requirements.txt
alembic
argparse
Flask
flask-cors
flask-login
Flask-WTF
googleapis-common-protos
grpcio
grpcio-tools
importlib_resources==1.4.0
mypy-protobuf
protobuf
psycopg2
pyzmq
requests
SQLAlchemy
squeaklib==0.5.3
typed-config
